// src/auth/role.enum.ts
export enum Role {
    SUPER_ADMIN = 'SUPER_ADMIN',
    INVENTORY_MANAGER = 'INVENTORY_MANAGER',
    SALES_MANAGER = 'SALES_MANAGER',
    ACCOUNTING_MANAGER = 'ACCOUNTING_MANAGER',
    CUSTOMER_SERVICE = 'CUSTOMER_SERVICE',
    MARKETING_MANAGER = 'MARKETING_MANAGER',
    SYSTEM_ADMIN = 'SYSTEM_ADMIN',
    WAREHOUSE_MANAGER = 'WAREHOUSE_MANAGER',
    SALES_REP = 'SALES_REP',
    ACCOUNTING_CLERK = 'ACCOUNTING_CLERK',
    CUSTOMER_SERVICE_REP = 'CUSTOMER_SERVICE_REP',
    MARKETING_ASSOCIATE = 'MARKETING_ASSOCIATE',
    SYSTEM_ANALYST = 'SYSTEM_ANALYST',
    WAREHOUSE_ASSOCIATE = 'WAREHOUSE_ASSOCIATE',
    // Add more roles as needed
    // For example, you might have roles like:Auditor, HR_Manager, etc.
    // HR_MANAGER = 'HR_MANAGER',
    // AUDITOR = 'AUDITOR',
}